chore: remove theme modules from Flow
Remove the Lumo and Material theme
modules from Flow project to have them
in the components project instead.

Closes #10182
