+++ .gitignore
@@ -57,3 +57,4 @@ flow-tests/**/types.d.ts
 /flow-tests/**/*.generated.js
 /flow-tests/**/generated/**
 flow-tests/**/pnpmfile.js
+flow-tests/*/*.npmrc
