+++ flow-tests/pom.xml
@@ -331,8 +331,6 @@
                 <module>test-servlet</module>
                 <module>test-application-theme</module>
                 <module>test-themes</module>
-                <module>test-lumo-theme</module>
-                <module>test-material-theme</module>
                 <module>servlet-containers</module>
 
                 <!-- web component embedding -->
