+++ flow-tests/test-embedding/test-embedding-theme-variant/src/test/java/com/vaadin/flow/webcomponent/ThemedVariantComponentIT.java
@@ -36,12 +36,5 @@ public class ThemedVariantComponentIT extends ChromeBrowserTest {
         TestBenchElement webComponent = $("themed-variant-web-component")
                 .first();
         Assert.assertEquals("dark", webComponent.getAttribute("theme"));
-
-        String customStyle = $("custom-style").first()
-                .getAttribute("innerHTML");
-        Assert.assertThat(customStyle,
-                CoreMatchers.allOf(
-                        CoreMatchers.containsString("theme~=\"dark\""),
-                        CoreMatchers.containsString("--lumo-base-color")));
     }
 }
