+++ flow-tests/test-live-reload/pom.xml
@@ -38,11 +38,6 @@
             <version>3.0.1</version>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>com.vaadin</groupId>
-            <artifactId>vaadin-lumo-theme</artifactId>
-            <version>${project.version}</version>
-        </dependency>
     </dependencies>
 
     <build>
