--- flow-tests/test-lumo-theme/frontend/theme/lumo/LumoThemedTemplate.js
@@ -1,23 +1,0 @@
-import {PolymerElement} from '@polymer/polymer/polymer-element.js';
-import {html} from '@polymer/polymer/lib/utils/html-tag.js';
-
-class LumoThemedTemplate extends PolymerElement {
-  static get is() { return 'explicit-lumo-themed-template' }
-
-  static get template() {
-    return html`
-       <div id='div'>Lumo themed Template</div>
-       <style>
-       div {
-        color: var(--lumo-error-color); /* color */
-        font-size: var(--lumo-font-size-xxxl); /* typography */
-        border: var(--lumo-size-m) solid black; /* sizing */
-        margin: var(--lumo-space-wide-l); /* spacing */
-        border-radius: var(--lumo-border-radius-l); /* style */
-        font-family: lumo-icons; /* icons */
-       }
-       </style>
-    `;
-  }
-}
-customElements.define(LumoThemedTemplate.is, LumoThemedTemplate);
