--- flow-tests/test-material-theme/src/test/java/com/vaadin/flow/uitest/ui/material/NotThemedTemplateIT.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2000-2020 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.uitest.ui.material;
-
-import com.vaadin.flow.testutil.ChromeBrowserTest;
-import com.vaadin.testbench.TestBenchElement;
-import org.junit.Assert;
-import org.junit.Test;
-
-public class NotThemedTemplateIT extends ChromeBrowserTest {
-
-    @Test
-    public void materialThemeUsed_noMaterialTemplateFile_srcBasedTemplateUsed() {
-        open();
-
-        // check that all imported templates are available in the DOM
-        TestBenchElement template = $("not-themed-template").first();
-
-        TestBenchElement div = template.$("div").first();
-
-        Assert.assertEquals("Template", div.getText());
-    }
-
-}
