--- flow-theme-integrations/lumo/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-SymbolicName: ${project.groupId}.flow.theme.lumo
-Bundle-Name: Vaadin Flow Lumo Theme
-Bundle-Version: ${osgi.bundle.version}
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-License: http://www.apache.org/licenses/LICENSE-2.0
-Export-Package: com.vaadin.flow.theme.lumo*;-noimport:=true
