--- flow-theme-integrations/lumo/src/test/com/vaadin/flow/theme/lumo/LumoTest.java
@@ -1,5 +1,0 @@
-package com.vaadin.flow.theme.lumo;
-
-public class LumoTest {
-
-}
