--- flow-theme-integrations/material/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-SymbolicName: ${project.groupId}.flow.theme.material
-Bundle-Name: Vaadin Flow Material Theme
-Bundle-Version: ${osgi.bundle.version}
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-License: http://www.apache.org/licenses/LICENSE-2.0
-Export-Package: com.vaadin.flow.theme.material*;-noimport:=true
